Move module augmentation to the index
